Updated from Charlies Local folders

+++ BrainSimulator/FiringHistoryWindow.xaml.cs
@@ -40,10 +40,9 @@ namespace BrainSimulator
         public void Draw()
         {
             //if you are not full-width or at end, don't repaint because it makes the image move around
-            bool atRightEnd = (int)(scroller.HorizontalOffset + scroller.RenderSize.Width) == (int)scroller.ExtentWidth;
+            bool atRightEnd = scroller.HorizontalOffset + scroller.RenderSize.Width == scroller.ExtentWidth;
             bool fullWidth = scroller.RenderSize.Width == scroller.ExtentWidth;
-            if (theCanvas.Children.Count != 0 && !atRightEnd && !fullWidth) 
-                return;
+            if (theCanvas.Children.Count != 0 && !atRightEnd && !fullWidth) return;
             if (dragging) return;
             ReallyDraw();
         }
